requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.4
idna==2.6
python-crontab==2.2.4
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
six==1.10.0
urllib3==1.22
